+++ src/format.jl
@@ -1,14 +1,37 @@
 using Dates: Dates
 
+# fast number to string helpers
+@inline function _write2!(p::Ptr{UInt8}, x::UInt)  # 00 - 99
+    unsafe_store!(p, 0x30 + (x ÷ 10))      # ‐‐> ‘0’ + tens
+    unsafe_store!(p + 1, 0x30 + (x % 10))  # ‐‐> ‘0’ + ones
+end
+
+@inline function _write4!(p::Ptr{UInt8}, x::UInt)  # 0000 - 9999
+    unsafe_store!(p, 0x30 + (x ÷ 1000))
+    unsafe_store!(p + 1, 0x30 + (x ÷ 100 % 10))
+    unsafe_store!(p + 2, 0x30 + (x ÷ 10 % 10))
+    unsafe_store!(p + 3, 0x30 + x % 10)
+end
+
+@inline function _write9!(p::Ptr{UInt8}, x::UInt)  # 000_000_000 - 999_999_999
+    @inbounds for k in 8:-1:0
+        d = x % 10
+        unsafe_store!(p + k, 0x30 + d)
+        x ÷= 10
+    end
+    nothing
+end
+
 """
     ISOTimestamp64Format
 
-Describes the ISO 8601 formatting for a `Timestamp64`.
+Describes the ISO 8601 nano formatting for a `Timestamp64`.
 This is the default value for `Dates.format` of a `Timestamp64`.
+Note that it does not contain a trailing `Z` character.
 
 # Example
 ```jldoctest
-julia> Dates.format(Timestamp64(2018, 8, 8, 12, 0, 43, 1), ISOTimestamp64Format)
+julia> Dates.format(Timestamp64(2018, 8, 8, 12, 0, 43, nanoseconds=1), ISOTimestamp64Format)
 "2018-08-08T12:00:43.000000001"
 ```
 """
@@ -16,14 +39,27 @@ const ISOTimestamp64Format = Dates.DateFormat("yyyy-mm-dd\\THH:MM:SS.fffffffff")
 Dates.default_format(::Type{Timestamp64}) = ISOTimestamp64Format
 
 """
+    RFC3339Timestamp64Format
+
+Describes the RFC 3339 nano formatting for a `Timestamp64`.
+This is the default value for `Dates.format` of a `Timestamp64`.
+Note that it does not contain a trailing `Z` character.
+
+# Example
+```jldoctest
+julia> Dates.format(Timestamp64(2018, 8, 8, 12, 0, 43, nanoseconds=1), ISOTimestamp64Format)
+"2018-08-08T12:00:43.000000001"
+```
+"""
+const RFC3339Timestamp64Format = Dates.DateFormat("yyyy-mm-dd\\THH:MM:SS.fffffffffZ")
+
+"""
     Dates.format(t::Timestamp, ::typeof(ISOTimestamp64Format)
 
 Return a string representing the timestamp in ISO 8601 format
 with nanosecond precision.
 """
-function Dates.format(timestamp::Timestamp64, ::typeof(ISOTimestamp64Format))
-    iso8601(timestamp)
-end
+Dates.format(timestamp::Timestamp64, ::typeof(ISOTimestamp64Format)) = iso8601(timestamp)
 
 """
     Dates.format(t::Timestamp, ::typeof(Dates.ISODateTimeFormat)
@@ -31,95 +67,88 @@ end
 Return a string representing the timestamp in ISO 8601 format
 with nanosecond precision.
 """
-function Dates.format(timestamp::Timestamp64, ::typeof(Dates.ISODateTimeFormat))
-    iso8601(timestamp)
-end
+Dates.format(timestamp::Timestamp64, ::typeof(Dates.ISODateTimeFormat)) = iso8601(timestamp)
+
+"""
+    Dates.format(t::Timestamp, ::typeof(RFC3339Timestamp64Format)
+
+Return a string representing the timestamp in RFC 3339 format
+with nanosecond precision.
+"""
+Dates.format(timestamp::Timestamp64, ::typeof(RFC3339Timestamp64Format)) = rfc3339(timestamp)
 
 """
     iso8601(timestamp::Timestamp64)
 
-Return a string representing the timestamp in ISO 8601 format.
+Return a string representing the timestamp in ISO 8601 nano format.
 
-Note that the timestamp is always in UTC and the precision is nanoseconds (9 digits).
+Note that the timestamp is always in UTC,
+the precision is nanoseconds (9 digits),
+and **no** trailing 'Z' character is returned.
 
 # Examples
 
 ```jldoctest
-julia> using Timestamps
-
+julia> iso8601(Timestamp64(2018, 8, 8, 12, 0, 43, nanoseconds=1))
+"2018-08-08T12:00:43.000000001"
 ```
 """
-function iso8601(timestamp::Timestamp64)
-    time = timestamp.ts
-    out = "0000-00-00T00:00:00.000000000"
-    ptr = pointer(out)
-    nss = 1_000_000_000
+function iso8601(ts::Timestamp64)
+    base = "0000-00-00T00:00:00.000000000" # 29 chars
 
-    z = time ÷ (86_400 * nss) + 719468
-    era = div(z >= 0 ? z : z - 146096, 146097)
-    doe = z - era * 146097
-    yoe = div(doe - div(doe, 1460) + div(doe, 36524) - div(doe, 146096), 365)
-    year = yoe + era * 400
-    doy = doe - (365 * yoe + div(yoe, 4) - div(yoe, 100))
-    mp = div(5 * doy + 2, 153)
-    day = doy - div(153 * mp + 2, 5) + 1
-    month = mp + (mp < 10 ? 3 : -9)
-    year += (month <= 2)
+    GC.@preserve base begin
+        ptr = pointer(base)
 
-    # TODO: Optimize
-    date_str = string(1_000_000 * year + 1000 * month + day)
-    unsafe_copyto!(ptr, pointer(date_str), 10)
+        # write date
+        y, m, d = Dates.yearmonthday(ts)
+        _write4!(ptr, UInt(y))
+        _write2!(ptr + 5, UInt(m))
+        _write2!(ptr + 8, UInt(d))
 
-    hour = (time ÷ (3_600 * nss)) % 24
-    mins = (time ÷ (60 * nss)) % 60
-    secs = (time ÷ nss) % 60
-    ns = (time % nss)
+        # write time
+        _write2!(ptr + 11, UInt(Dates.hour(ts)))
+        _write2!(ptr + 14, UInt(Dates.minute(ts)))
+        _write2!(ptr + 17, UInt(Dates.second(ts)))
+        _write9!(ptr + 20, UInt(Dates.nanosecond(ts)))
+    end
 
-    t = (100_000_000 * nss + hour * 10_000_000 * nss + mins * 10_000 * nss + secs * 10 * nss + ns)
-    # TODO: Optimize
-    t_str = string(t)
-    unsafe_copyto!(ptr + 11, pointer(t_str), 18)
-
-    unsafe_store!(ptr, '-', 5)
-    unsafe_store!(ptr, '-', 8)
-    unsafe_store!(ptr, ':', 14)
-    unsafe_store!(ptr, ':', 17)
-    unsafe_store!(ptr, '.', 20)
-    unsafe_store!(ptr, unsafe_load(ptr, 12) - 1, 12)
-
-    out
+    base
 end
 
-## Does the same, but is slower
-# function iso8601(timestamp::Timestamp64)
-#     time = timestamp.ts
-#     nss = 1_000_000_000
-#     # mn = 1_000_000
+"""
+    rfc3339(timestamp::Timestamp64)
 
-#     z = time ÷ (86_400 * nss) + 719468
-#     era = div(z >= 0 ? z : z - 146096, 146097)
-#     doe = z - era * 146097
-#     yoe = div(doe - div(doe, 1460) + div(doe, 36524) - div(doe, 146096), 365)
-#     year = yoe + era * 400
-#     doy = doe - (365 * yoe + div(yoe, 4) - div(yoe, 100))
-#     mp = div(5 * doy + 2, 153)
-#     day = doy - div(153 * mp + 2, 5) + 1
-#     month = mp + (mp < 10 ? 3 : -9)
-#     year += (month <= 2)
+Return a string representing the timestamp in RFC 3339 nano format.
 
-#     year_str = string(year)
-#     month_str = lpad(month, 2, '0')
-#     day_str = lpad(day, 2, '0')
+Note that the timestamp is always in UTC,
+the precision is nanoseconds (9 digits),
+and a trailing 'Z' character is returned.
 
-#     hour = (time ÷ (3_600 * nss)) % 24
-#     mins = (time ÷ (60 * nss)) % 60
-#     secs = (time ÷ nss) % 60
-#     ns = (time % nss)
+# Examples
 
-#     hour_str = lpad(hour, 2, '0')
-#     mins_str = lpad(mins, 2, '0')
-#     secs_str = lpad(secs, 2, '0')
-#     ns_str = lpad(ns, 9, '0')
+```jldoctest
+julia> rfc3339(Timestamp64(2018, 8, 8, 12, 0, 43, nanoseconds=1))
+"2018-08-08T12:00:43.000000001Z"
+```
+"""
+function rfc3339(ts::Timestamp64)
+    base = "0000-00-00T00:00:00.000000000Z" # 30 chars
 
-#     "$(year_str)-$(month_str)-$(day_str)T$(hour_str):$(mins_str):$(secs_str).$(ns_str)Z"
-# end
+    GC.@preserve base begin
+        ptr = pointer(base)
+
+        # write date
+        y, m, d = Dates.yearmonthday(ts)
+        _write4!(ptr, UInt(y))
+        _write2!(ptr + 5, UInt(m))
+        _write2!(ptr + 8, UInt(d))
+
+        # write time
+        _write2!(ptr + 11, UInt(Dates.hour(ts)))
+        _write2!(ptr + 14, UInt(Dates.minute(ts)))
+        _write2!(ptr + 17, UInt(Dates.second(ts)))
+        _write9!(ptr + 20, UInt(Dates.nanosecond(ts)))
+    end
+
+    base
+end